blogisofta/requirements.txt
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pygame==2.0.1
Werkzeug==2.0.1
psycopg2==2.9.2
google-cloud-secret-manager==2.0.0